PermutationContainer: Fix undefined behavior
When using a pointers of a different type (Permutation instead of e.g.
ExplicitPermutation), they will be incremented by the wrong amount
(sizeof(Permutation) instead of sizeof(ExplicitPermutation)) and thus
the pointers end up pointing to random memory locations which will cause
a segfault, when dereferenced.

Instead a custom iterator implementation must be used that explicitly
stores the underlying type's size and accounts for it accordingly.