Fix build as sub-project
CMAKE_SOURCE_DIR will always point to the top-level project's root
source directory, which is not what we want. Instead, we now use
PROJECT_SOURCE_DIR that will ensure that we will always fetch the source
root of libPerm itself.
(Otherwise the include paths for libPerm won't be resolved correctly)