Add support for interfacing perms and containers
New functions have been added, that allow creating a permutation that,
applied to a given container, will bring its elements into sorted order,
based on the default (less-than) or provided comparator. Additionally,
a function was added that can then actually apply a given Permutation to
the elements of a container, in order to permute them in-place.

Note: Supported containers must provide random-access to their elements.