Fix issues with computing sort & applying permutations
The meaning of these permutations was reversed from what was originally
intended.
The intended meaning is for the permutation to give the new location of
an element at position i (that is i^p is the new position, if p is the
permtuation in question).
However, the meaning that was actually encoded in the implementation was
that the permutation describes what element of the original container
shall be positioned at the ith position of the sorted one (That is
container[i^p] is the element that shall be placed at the ith position
in the sorted container).

These differences are rather subtle and only matter for cyclic
permutations of more than three elements.