Make inverting sign optional
When inverting a given permutation, inverting the sign along with it, is
not always desirable. While it is necessary to strictly obtain
g * g^{-1} = id
for signed permutations, it gets in the way when working in contexts,
where a permutation's sign doesn't have any meaning (arguably, in those
cases one shouldn't use signed permutations to begin with, but the
library doesn't really support that yet).