Support pre-multiplication of permutations
Previously, permutations only allowed for post-multiplication, meaning
that a.mulitply(b) resulted in a * b. There was no way to instead
calculate b * a, if one wanted the modified result to be contained in a.

This change deprecates the old AbstractPermutation::mulitply method in
favor of the more explicit preMultiply and postMultiply methods.

Additionally, with these methods available, the binary multiplication
operator of ExplicitPermutation can now take an arbitrary permutation as
either the lhs or the rhs, making it more flexible.