Support computation of left and right coset (representatives)
Instead of always only considering the right coset when determining the
coset representative, we now add support for left and right cosets.
The old function to get the coset representative (which doesn't specify
which coset to consider) is now deprecated and defaults (as before) to
the right coset.

Additionally, functions to explicitly compute and return the left and
right coset have been added.