Change interpretation of perm acting on sequence
We changed the interpretation from applying the permutation p to the
sequence s from
new[i^p] = s[i]
to
new[i] = s[i^p]
which has the benefit that applying a permutation to a sequence of
increasing integers, starting at zero, will yield the permutation's
image rather than a (seemingly) arbitrary order of the integers.

Note, however, that in the new interpretation the order of composition
reverses when applying permutations to lists rather than integers. While
the permutation c = a * b, means "first apply a, then apply b" when
applying c to an integer (left-to-right), that same permutation means
"first apply b, then apply a" when applying to a sequence
(right-to-left).